+++ src/libraries/System.IO.Pipelines/src/System/IO/Pipelines/BufferSegment.cs
@@ -9,7 +9,7 @@ using System.Threading;
 
 namespace System.IO.Pipelines
 {
-    internal sealed class BufferSegment : ReadOnlySequenceSegment<byte>
+    internal sealed partial class BufferSegment : ReadOnlySequenceSegment<byte>
     {
         private object? _memoryOwner;
         private BufferSegment? _next;
@@ -58,33 +58,6 @@ namespace System.IO.Pipelines
         {
             _memoryOwner = arrayPoolBuffer;
             AvailableMemory = arrayPoolBuffer;
-        }
-
-        public void ResetMemory()
-        {
-            object? memoryOwner = Interlocked.Exchange(ref _memoryOwner, null);
-            if (memoryOwner is null)
-            {
-                // Already returned block; this can happen from benign race between Advance and Complete
-                // as block return is done outside a lock.
-                return;
-            }
-
-            if (memoryOwner is IMemoryOwner<byte> owner)
-            {
-                owner.Dispose();
-            }
-            else
-            {
-                Debug.Assert(memoryOwner is byte[]);
-                byte[] poolArray = (byte[])memoryOwner;
-                ArrayPool<byte>.Shared.Return(poolArray);
-            }
-
-            RunningIndex = 0;
-            Memory = default;
-            _end = 0;
-            AvailableMemory = default;
         }
 
         public void ResetSegmentLinks()
